Still work in progress -- added the first cut of the model for the water feeds from the manifold to the boards.
The circular distribution pipes will need to be modeled differently
in the next cut as the simple circular description doesn't fir too
well with various other volumes around.

Modified the copper cable bundle geometry to allow for the inclusion
of the water feeds. Chnaged the strut material from Be to Carbon.