Revert "Remove throw"
No value can be immediately seen in the changes.
We need to hear more arguments on why it is better to remove the throw. What is
the point in adding a dummy static local variable that does nothing?
Even if some changes can be defended to be appropriate the same must be done in
StiIstHitLoader and StiSsdHitLoader