Version vetted by UPC
StPPVertexFinder: Setup fit with single parameter for 1D fit

StPPVertexFinder: Changed requirements for PPV vertex fit

For vertex reconstruction require at least two track or a track and a beam line
constraint

StPPVertexFinder: Track's mother vertex id can be negative

Negative id's are used to indicate non-triggered vertices but the
mother-daughter relation between the vertex and a track is still valid

StPPVertexFinder: Keep vertices with failed fits

This does not change the previous logic. However, the vertex position and errors
are set to that of the beam line. It might make sense to save the fit status in
the vertex object later.

StPPVertexFinder: Remove vertex seeds which cannot be fit