The geometry is steered by a set of user commands, input using the DETP facility, e.g.
   DETP GEOM y2014a field=-5.0 gcalor_on=1 ...

In AgML 1, the list of commands was iterated over.  Each command which was
recognized resulted in several variables being set, which would steer the
execution of the geometry modules at a later point in the code.

Prior to executing the modules, a volume was defined where secondary
tracks would be saved.

With AgML 2, the geometry tag is executed during the loop as soon as it is
recognized.  Thus, the definition of the volume where secondardies are
saved has to be moved earlier in the code.