The order in which the AgSFlag and AgDETP commands are executed matters.
AgDETP NEW creates and initializes the data structures for a new detector
module.  It is responsible for picking up the flags specified for assigning
hits from secondary tracks.  By default, this is set "calorimeter" style
assignment.  Hits on secondary tracks are assigned to the track which
initiated the decay and/or shower.

The order in which these commands are executed therefore matters.  Flags
set after the initialization of the data structure will not be copied into
the current data structure... but rather the next one.

AgML 1 handled this properly.

AgML 2 inadvertantly changed the order.  Thus, tracking detectors initialized
after a calorimeter-like detector picked up the calorimter style hit assignment,
and vice versa.

And so, see trouble ticket 3271.

This commit fixes the order in AgML 2.  AgSFlag is called before the detector
data structures are initialized.