[www.star.bnl.gov #3364] Distance Units in the Simulation Framework
Issue with the system of units.  StarGenEvent records distance in units
of mm, consistent with the HEPEVT standard used in event generators.

(All of our generators, except for the cosmic-ray option in Kinematics,
respect this standard).

The StarParticleStack assumes distances to be measured in cm, consistent
with the GEANT3 system of units.

StarPrimaryMaker was supposed to make the conversion, but this was not coded.

Consequently, decays of unstable particles were being pushed to GEANT
with an incorrect vertex.

Primary particles pushed out at z = 0, so not an issue for those.

The fix is to bring StarPrimaryMaker into compliance with the HEPEVT standard.

Previous productions are largely unaffected, because we flag (most) particles
as stable at the event generation stage, and allow geant to perform decays.
Once the particle is on the geant stack, distances are handled in a consistent
set of units.  Embedding productions have typically not been done in this
framework, so are not affected by this bug.  Any use of the Kinematics
"particle gun" event generator is also unaffected.