Implemented a templated method for creating and filling the g2t hit tables.
This reduces the boilerplate code required when integrating a new detector
subsystem, and streamlines maintanence.

We also add a summary printout of the number of hits found in each subsystem,
allowing nightly tests to keep track of changes in the geometry introduced
at the hit level.

Code was tested for the first cut and last production geometry in each run
from 2005 to 2018.  For 1k pythia event -4.5 < eta < 4.5, geant.root file
size is nearly identical (~few to 50 bits/event) comparing old and new version
of St_geant_Maker.  When file compression is switched off, file size is identical.