The incoming version of the FTS geometry takes ~30min to convert from AgML to the compilable "cxx" files, signifincatly longer than other geometries.  Conversion to "age" files does not take notiably longer than other geometries.
Running the python profiler on AgML identifies the "replacements()"
function as the leading time consumer.  This function performs a
number of text searches using python's regular expression module.
Several regex's are processed on each line of XML code which agml
parses.  The result is a significant slow down as some fairly complicated
regexs must be compiled over and over again.

This commit adds two guard statements, which perform a fast check to
see if the line in question needs to be tested against the expensive
regular expressions.

To verify that these changes have no impact on the code, I performed
a unix diff on the ".cxx" and ".h" outputs, comparing before and after
this change.  There were zero differences found.