Final commit to refactor StarGenerator, in order to move the dependency on the starsim libraries out of the event generator framework.
The AgStarReader framework  is moved into St_geant_Maker.  Initialization
of the reader presents a small problem.  Most of our example macros have
delayed initialization of the generator framework until *after* the
initialization of the full chain.  So we don't have a guarenteed hook
to setup the connection between the event generator and the starsim
package.  Thus, the only place we can do this is on the first call to
St_geant_Maker::Make.  It is somewaht ugly, but the alternative is to
change the initialization scheme, change all of the production and example
macros, AND handle the fallout of breaking *users* macros.