Issue #45 notes an issue with building the geometry codes.  This is confirmed with a fresh clone, and execution of cons.
The dependencies for building the geometry modules are set as:

- module.o: module.h module.cxx
- module.h: module.xml agmlParser.py
- module.cxx: module.xml agmlParser.py

Removing the dependency on the parser allows a freshly cloned repo to compile,
and may resolve issue #45.  It is not clear to me why.

Note--

Originally, the parser was added to the dependency list so that when we
changed the agml compiler, we would automatically rebuild the geometry.
This is no longer necessary.  AgML is fairly stable.  Plus, any changes to
the language will be accompanied by changes in the module.xml files.

We drop this dependency with this commit.