The libTable library was deprecated following ROOT 6.16.  In order to support our framework with later versions of ROOT, we (re) import the libTable sources into our repository.
In order to support building our software stack with versions of ROOT which may
provide support for libTable, we modify cons to detect whether or not libTable
support was compiled in.  If ROOT provides libTable, we do not build it and
do not export header files into $STAR_HOST_SYS/include.  If ROOT does not provide
libTable, we build it ourselves.