starsim: make seed overflow into a fatal error
The value of bigger than 0x7FFFFFFF will overflow ISEQ and make it
wrap around into a negative values. The current behavior is to print a
warning message and reset the seed to a default value. Because of that
an unsuspecting user can easily initialize a fraction of their
simulation production with the same seed.