Refactor AgML volume extensions to utilize a map (global, static, ugly) instead of hanging themselves off of the geometry tree.
The GEANT3 VMC implementation of sensitive volumes makes (inappropriate)
use of the "user" extension attached to geometry volumes.  This requires
us to attach extended information using a separate data structure,
otherwise it will be overwritten.

This is the first stage of refactoring, adding the volume extension map
to the AgML extension class as a static object.