Setup code to enable standalone simulation with GEANT3, Geant 4 or mixed simulation modes.
Requires that geant321 option is provided on the command line (loading
the GEANT3 library).

At present, the GEANT3 standalone code initializes and configures, but
produces no hits.  Reason TBD.

At present, the code needed to switch a track between physics engines
has not been implemented.  So only the standalone modes are supported.
In order to implement, we need to add code to the stepping routines
which shunt the track between engines.  We will likely also need to
keep track of which engine a track is currently in so that we are not
constantly switching for no reason.