When using multi-engine mode, the stack is managed by the TMCManager. The StMCParticleStack (user stack) is used as input, and held by the TMCManager.  The TGeant3 and TGeant4 MC engines hold pointers to two stacks (TMCManagerStack), rather than the user stack.
The StMCParticleStack contains the truth table, allowing us to associate
hits from primary and secondary to their tracks, and hits from showering
particles to the primary/secondary which initiated the shower.

Access to the StMCParticleStack cannot be guarenteed through the pointers to
the MC engine.  Thus, we need to modify the hit collections / sensitive detectors
to directly hold a pointer to the "user" stack.