Test ID truth for single engine mode (or for that matter, multi engine mode...)
Throws N particles w/in the BEMC acceptance, and requires that all
primary tracks register hits.  (Don't throw neutrinos, obviously...)

We have corret behavior for single G3, single G4 modes, and multi-engine
modes where tracks are never transfered between engines.

----

When a track is transfered between engines, it is moved from the
current stack to the new engine stack and the track is interupted.
Once all tracks on the current stack have been processed, the new
engine will process tracks on its stack in the order in which they
were transfered (or reverse order?).

This breaks one of the assumptions of the idtruth following scheme.
Namely, that a track is processed until it interacts, decays or
otherwise is stopped.

So... a redesign of the ID truth / hit scoring scheme is in order
for the case where we have a multi-engine run.