bfc.C: fix for cling
cling needs to have libraries loaded before running a macro.

This disables the #include statements when ACLiC is not
used. Currently, cling can't parse those headers: there seems to be an
issue with how forward declarations interract with
"$clingAutoload$". Fortunately, the "#pragma load" takes care of the
headers, like in ROOT5, so things just work.