StGenericVertexFinder: Avoid floating point precision caramba
It appears that the error transformation may break when we get closer to
the nominal beam position. In other words, all the Jacobian components
can become 0 if not enough precision is used to calculate them. This
makes the further calculations pointless by yielding a NaN (0/0). We
avoid this by doing a reqularization on an irreducible beam width on the
order of a nanometer.