Stabilize calculation of uncertainties for vertices extremely close to beam line (#279)
* fix an assertion failure in 64bit Run22 pp510 data production. The fix comes from Yuri Fisyak.

* StGenericVertexFinder: Avoid floating point precision caramba

It appears that the error transformation may break when we get closer to
the nominal beam position. In other words, all the Jacobian components
can become 0 if not enough precision is used to calculate them. This
makes the further calculations pointless by yielding a NaN (0/0). We
avoid this by doing a reqularization on an irreducible beam width on the
order of a nanometer.

* Update StRoot/StGenericVertexMaker/StGenericVertexFinder.cxx

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>

Co-authored-by: Dmitri Smirnov <dmixsmi@gmail.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>