Fix static linking of mysql library to root4star
According to cons scripts, the entire `mysqlclient` library is requested
to be linked statically to `root4star` using the following linker flags:

```
...
-Wl,--whole-archive -Wl,-Bstatic -Wl,-z -Wl,muldefs
`mysql_config --libs`
-Wl,--no-whole-archive -Wl,-Bdynamic
...
```

The linker is instructed to look for static versions of the libraries
enclosed between the `-Wl,--whole-archive` and `-Wl,--no-whole-archive`
options, however when such libs cannot be found it rightfully fails.
Now, normally, `mysql_config --libs` is supposed to return something
similar to this:

```
-L$pkglibdir -lmysqlclient -lpthread -lz -lm -lssl -lcrypto -ldl
```

But the `mysql_config` script on the farm was apparently modified to
match the specific linker command in cons, so it returns additional
flags:

```diff
$ diff /opt/star/sl73_gcc485/bin/mysql_config /opt/star/sl73_gcc485/bin/mysql_config~
110c110
< libs=" $ldflags -L$pkglibdir  -lmysqlclient   -L/usr/lib -Wl,--no-whole-archive -Wl,-Bdynamic -lpthread -lz -lm -lssl -lcrypto -ldl "
---
> libs=" $ldflags -L$pkglibdir  -lmysqlclient   -lpthread -lz -lm -lssl -lcrypto -ldl "
112c112
< libs_r=" $ldflags -L$pkglibdir   -lmysqlclient  -L/usr/lib -Wl,--no-whole-archive -Wl,-Bdynamic  -lpthread -lz -lm -lssl -lcrypto -ldl   "
---
> libs_r=" $ldflags -L$pkglibdir   -lmysqlclient   -lpthread -lz -lm -lssl -lcrypto -ldl   "
```

```shell
$ ls -la /opt/star/sl73_gcc485/bin/mysql_config /opt/star/sl73_gcc485/bin/mysql_config~
-rwxr-xr-x 1 jeromel rhstar 6842 Oct 10  2017 /opt/star/sl73_gcc485/bin/mysql_config
-rwxr-xr-x 1 jeromel rhstar 6746 Oct 10  2017 /opt/star/sl73_gcc485/bin/mysql_config~
```

The point of this manual intervention was obviously to exclude the
non-existing static libraries following the `-Wl,-Bdynamic` option.
Since modifying generated package configurations is never a good idea,
this fix will allow to rely on standard mysql installations without
any further patching of cons scripts.