bfc.C: fix for cling (#278)
* bfc.C: fix for cling

cling needs to have libraries loaded before running a macro.

This disables the #include statements when ACLiC is not
used. Currently, cling can't parse those headers: there seems to be an
issue with how forward declarations interract with
"$clingAutoload$". Fortunately, the "#pragma load" takes care of the
headers, like in ROOT5, so things just work.

* bfc.C: add more libraries needed for ROOT 6.16

When tested on 6.24/06 the libraries seem to be loaded
automagically. On CI we use 6.16, and that seems to not work. Or it
could be that it uses RTLD_NOW. Let's try to add missing libraries
manually.