ci: avoid a duplicate step in build-containers.yml
Originally, the "latest" containers for a branch were supposed to
provide a cache for the build environment for the PRs and new
tags. Since the switch to the new star-spack infrastructure the
environment is now built elsewhere, the only target left in the local
dockerfile is to build the star-sw itself.