cons: Help agetof to find definition file (#381)
* starsim: Do not overwrite global agetof options

* cons: Help agetof to find definition file

When upgrading Spack (used to build the STAR environment in containers)
from v0.17.2 to v0.18.1, building with `cons` fails due to an issue with
the `agetof` utility. It seems to have trouble finding the definition
file `asps/Simulation/agetof/agetof.def` normally installed in
`$STAR_BIN` along with the `agetof` executable.

The fix appears to be backward compatible and should work for all
versions of `agetof` accepting the argument.