When we try to run bfc.C using ROOT6 / cling... cling complains about several things being redefined.  It points to this BFC.C macro in StBFChain/.
BFC.C is loaded and executed through the ROOT interpreter in StBFChain::Setup.

Minimal solution is to protect the table implementation for "Bfc" from being
compiled by the cling interpreter (just like it is protected from compilation
by the CINT interpreter).

Likely solves the same problem which Victor was addressing on the Star6
branch.
https://github.com/star-bnl/star-sw/blame/Star6/StRoot/StBFChain/BFC.C#L6