[root6] StEvent/StEnumerations.cxx fix (#440)
StEvent/StEnumerations.cxx uses a _ROOT macro_ (under
StRoot/macros/detectorId.C) to build a reflection map between the
StDetectorId enum and the name of each enum.

Under root 5, the current code works fine. But it is ugly. We should at
refactor the code to use a helper function compiled in the library. This
PR makes that update.

Under root 6, the current code fails. Because root 6 tries to compile
the macro, and the arguments to strncpy should not be const. Again, this
PR makes detectorId a helper function w/in StEnumerations.

---

We move the functionality provided by StRoot/macros/detectorId.C into
the compiled StEvent/StEnumerations.cxx file.

We retain StRoot/macros/detectorId.C in case one or more user macros
depend on it.

Co-authored-by: Dmitri Smirnov <dmixsmi@gmail.com>