[root6] Eliminate crash when creating the MuDst (#442)
When creating the MuDst, ROOT6 has a problem with creating the
TClonesArray
of StStrangeAssoc objects. Running under GDB shows that TClass::GetClass
enters
an infinite recursion... almost as if the system is confused, and
believes the class is it's
own base class.  

Disabling the call to initializeForWriting() on loading the share
library resolves the issue.
Though I don't know what side effects this may cause.

Co-authored-by: Dmitri Smirnov <dmixsmi@gmail.com>