Remove StBFChain dependence in StarDb/AgMLGeometry/CreateGeometry.h (#462)
This PR Should resolve #455.

The StBFChain dependence supported two features in this macro. 

1) It allowed us to name the geometry cache file based on the output
filename for the chain.
2) It allowed us to detect when Sti (CA or V) was present in the chain
options.

Item 1 is nice to have but not required.

Given the "assert(0)" in the block of code which detects Sti... we don't
actually use item 2. (I suspect it was deprecated at some point).

Therefore we remove the dependency on StBFChain.