[root6] Address issues in TTable when interpreted by Cling (#449)
Our ROOT6 tests indicated a problem with the ROOT's TTable class when it
is "interpreted" by the `bfc.C` macro. A few fixes seem to take care of that
but now we need to build and use the local code from `StRoot/Table`
instead of the external `libTable` provided by ROOT. This is achieved by
building `StRoot/Table` conditionally only when external dependency on
ROOT >= 6 is detected.