Fix errant switch-case enumeration (#577)
The cases in a switch statement went from 3 to 5, errantly skipping over
4. This patch affects no former productions as no detector > 3 has been
used before now. However, for FXT data, we will want detector > 3, so we
should fix this now for possible use in such productions.

It will also probably be worth patching this in SL23d (because this does
not impact any past productions), allowing us to use it before the next
major release step (e.g. SL23e).