Update StFstFastSimMaker to write hits into the StFstHitCollection (#700)
as is done for the real data reco chain. I continue to write hits into
the StRnDCollection, but this is no longer used by downstream code.

Also add a function to StFstHit to allow setting the disk, wedge, and
sensor for sim hits. Does not change the StFstHit data structure.