Zeros out the dead zones in the TPC (#723)
This PR removes the deadzone in the TPC. Responsibility for simulating
the deadzone is then shifted to the gating grid simulation.

This PR impacts all geometry tags that use the TpceGeo3a, 4, 5(*) and 6
models.

This PR should be tested prior to merging. Note that this implementation
would have the following consequences...
... the correct version of the TPC slow simulator must be used from this
point forward ...
... any backport of the tpc geometry into an earlier embedding library
requires backport of the slow simulator and vice versa ...

Should have a discussion about this before mergeing.

----
(*) No tag uses the TpceGeo5 model.  Listed here for completeness.