New FWD tracking mode based on FST-track finding (#571)
These updates provide a new FWD tracking mode in which the track finding
is one with FST information.
Tracks are then fit to valid Seeds and then (optionally) any FTT hits
are added to the track if found along the trajectory.

This mode is needed for any analysis of real data at this point. That
would include fast offline for ongoing data taking.

---------

Co-authored-by: Dmitri Smirnov <dmixsmi@gmail.com>
Co-authored-by: jdb12@rice.edu <jdb12@rice.edu>