GtHash should use unsigned int rather than long, as its representation does not change between 32bit and 64bit. (#745)
Thanks, @plexoos for demonstrating that it works for a third party. I have now found where I hadn't changed one line in my version of the code, and this fix works fine for me now too.

We can go ahead and merge (doing so now), but we still have to decide on which libraries we'll need to re-tag and re-build with this.