Make use of already designated topology bits (#773)
Several already-designated bits in StTrackTopologyMap have remained
unused (zeros). However, there are StTrack flags set in StiStEventFiller
for exactly these that can be used to set the bits in the topology map.
This does require swapping the order of setting the flags to precede
filling the topology maps in StiStEventFiller. This PR does nothing to
alter the intent of bits in the topology map, i.e. no change to any
structure in StEvent itself, but does add accessor functions.

The advantage of filling the track topology map over the track flags is
that it is propagated all the way to the PicoDst, while the track flags
are not, thereby giving PicoDst users a handle on this information about
the tracks without changing anything about the PicoDst structure.

I plan to follow this PR with another that makes use of currently
undesignated bits in the track topology map to include additional track
flags that may be useful to PicoDst users as well.