Additional flagging of possible pile-up (#774)
Following quickly on the heels of PR #773 , I have more modifications to
StTrackTopologyMap to propose...

A few bits were labeled as "reserved for future use"....well, there
isn't much future left to add new detectors to STAR here in the last
couple months of running. So I propose to target these for some of the
more useful binary track flags instead of leaving them as useless zeros
for eternity:

1. post-crossing tracks (definitely contain pile-up hits on the tracks)
2. tracks crossing the TPC central membrane (unlikely to be pile-up)

Both of these are rather TPC-centric, but I think this is worthwhile
utilization of the topology map without requiring any structural changes
to the PicoDsts in particular.